app/evals/pyproject.toml
[project]
name = "evals"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepeval>=3.8.9",
    "langchain-openai>=1.1.10",
    "psycopg2-binary>=2.9",
]
